Removed non-standard __GCC and use _GNUC_ instead.
When linking against `lib3MF` on linux (I guess it’s the same thing on
Mac OSX), we have to define __GCC (`-D__GCC`) so that we can include
lib3mf headers. That’s annoying and shouldn’t be the case. gcc and g++
define _GNUC_ exactly for that purpose.