NC-7659: Use libzip callback interface
Instead of copiiing the full input stream into memory.
This decreases the memory consumption of lib3MF on import by the size of
the compressed zip-file.
The speed of the import is higher for files with large attachments (e.g.
large textures, binary blobs) and not slower for "regular" XML content of
3MFs.