Precision of float output can now be changed (#169)
Writer-class gets new methods GetDecimalPrecision and SetDecimalPrecision which sets the number of decimal digits (after the decimal dot) that are serialized into a file.
Currently, only the 3mf-writer makes use of this. ASCII-STL-writer should respect this, too, at some point.

Verified that write performance is not degraded by this change.