Link against libzip/zlib when USE_INCLUDED_{LIBZIP,ZLIB}=OFF is used
lib3MF.so is not explicitly linked against libzip or zlib even when
USE_INCLUDED_LIBZIP=OFF or USE_INCLUDED_ZLIB=OFF are used. These
dependencies are only specified in the "Libs" section of the pkg-config
file. People trying to link the library directly without relying on the
pkg-config file will face errors like

/usr/bin/ld: lib3MF.so.1.8.1.0: undefined reference to `deflateInit2_'

This can be immediately observed when trying to compile lib3mf with

cmake -DUSE_INCLUDED_LIBZIP=OFF -DUSE_INCLUDED_ZLIB=OFF
      -DLIB3MF_TESTS=ON

The solution is to link lib3mf against all the necessary libraries so
that they can be pulled in automatically during linking. This allows to
move them from "Libs" to "Libs.private" in the pkg-config file since
they do not need to be specified explicitly any more. With this change,
compilation with the test suite as shown above works.

As another small improvement, use @LIBZIP_LDFLAGS@ and @ZLIB_LDFLAGS@
instead of statically specifying the required libraries. This guarantees
that only the actually necessary libraries (depending on the value of
USE_INCLUDED_...) are added as dependencies.